% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions.
}
\examples{
\dontrun{
ts <- microbenchmark(
rwC = rwMetropolisC(0.5, x0, N=1000),
rwR = rw.Metropolis(0.5, x0, N=1000))
print(summary(ts)[, c(1,3,5,6)])

}
}
