% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.R
\name{rw.Metropolis}
\alias{rw.Metropolis}
\title{A rw.Metropolis sampler using R}
\usage{
rw.Metropolis(sigma, x0, N = 1000)
}
\arguments{
\item{sigma}{the sigma}

\item{x0}{the random numbers}

\item{N}{the number of samples}
}
\value{
a random sample of size \code{n}
}
\description{
A rw.Metropolis sampler using R
}
\examples{
\dontrun{
sigma<-c(0.1,0.2)
x0<-200
rw.Metropolis(sigma[1], x0, N=1000)
}
}
